5-octobre/model/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tensorflow==2.15.0
numpy>=1.19.2
torch>=1.6.0
transformers>=4.33.0